Add ValueSource for default URl resolution
This will be created for the build and maintain configuration cache compatibility
as the values are not resolved during configuration time.

Add flag to set for integration test
