+++ src/main/java/com/bmuschko/gradle/docker/DockerExtension.java
@@ -1,11 +1,15 @@
 package com.bmuschko.gradle.docker;
 
 import com.bmuschko.gradle.docker.internal.DefaultDockerConfigResolver;
+import com.bmuschko.gradle.docker.internal.DefaultDockerUrlValueSource;
 import com.bmuschko.gradle.docker.internal.DockerConfigResolver;
 import org.gradle.api.Action;
+import org.gradle.api.XmlProvider;
 import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.Property;
+import org.gradle.api.provider.ProviderFactory;
+import org.gradle.api.provider.ValueSourceParameters;
 
 import java.io.File;
 
@@ -63,11 +67,11 @@ public class DockerExtension {
 
     private final DockerRegistryCredentials registryCredentials;
 
-    public DockerExtension(ObjectFactory objectFactory) {
+    public DockerExtension(ObjectFactory objectFactory, ProviderFactory providerFactory) {
         DockerConfigResolver dockerConfigResolver = new DefaultDockerConfigResolver();
 
         url = objectFactory.property(String.class);
-        url.convention(dockerConfigResolver.getDefaultDockerUrl());
+        url.convention(providerFactory.of(DefaultDockerUrlValueSource.class, noneValueSourceSpec -> {}));
         certPath = objectFactory.directoryProperty();
 
         File defaultDockerCert = dockerConfigResolver.getDefaultDockerCert();
