+++ src/main/java/com/bmuschko/gradle/docker/internal/DefaultDockerConfigResolver.java
@@ -7,38 +7,6 @@ import javax.annotation.Nullable;
 import java.io.File;
 
 public class DefaultDockerConfigResolver implements DockerConfigResolver {
-
-    private static final Logger logger = Logging.getLogger(DefaultDockerConfigResolver.class);
-
-    @Override
-    public String getDefaultDockerUrl() {
-        String dockerUrl = getEnv("DOCKER_HOST");
-        if (dockerUrl == null) {
-            if (OsUtils.isWindows()) {
-                if (isFileExists("\\\\.\\pipe\\docker_engine")) {
-                    dockerUrl = "npipe:////./pipe/docker_engine";
-                }
-            } else {
-                //Added for possibility to run integration tests, also makes it possible to override use of /var/run/
-                boolean skipVarRun = Boolean.valueOf(System.getProperty("com.bmuschko.gradle.docker.internal.DefaultDockerUrlValueSource.skipCheckOfVarRun", "false"));
-                // macOS or Linux
-                if (isFileExists("/var/run/docker.sock") && !skipVarRun) {
-                    dockerUrl = "unix:///var/run/docker.sock";
-                } else if (isFileExists(System.getProperty("user.home") + "/.docker/run/docker.sock")) {
-                    dockerUrl = "unix://" + System.getProperty("user.home") + "/.docker/run/docker.sock";
-                }
-            }
-
-
-            if (dockerUrl == null) {
-                dockerUrl = "tcp://127.0.0.1:2375";
-            }
-        }
-
-        logger.info("Default docker.url set to " + dockerUrl);
-        return dockerUrl;
-    }
-
     @Nullable
     @Override
     public File getDefaultDockerCert() {
