+++ src/main/java/com/bmuschko/gradle/docker/internal/DockerConfigResolver.java
@@ -3,7 +3,5 @@ package com.bmuschko.gradle.docker.internal;
 import java.io.File;
 
 public interface DockerConfigResolver {
-    String getDefaultDockerUrl();
-
     File getDefaultDockerCert();
 }
